+++ importers/gumroad/importer.js
@@ -1,16 +1,13 @@
+const cloudscraper = require('cloudscraper');
 const { posts, bans } = require('../../db');
 const scrapeIt = require('scrape-it');
-const Promise = require('bluebird');
-const fs = require('fs-extra');
-const crypto = require('crypto');
-const request = require('request');
-const request2 = require('request').defaults({ encoding: null });
-const { slugify } = require('transliteration');
-const indexer = require('../../indexer');
-const { URL } = require('url');
+const path = require('path');
 const checkForFlags = require('../../flagcheck');
-const cloudscraper = require('cloudscraper');
-const retry = require('p-retry');
+const downloadFile = require('../../download');
+const Promise = require('bluebird');
+const { URL } = require('url');
+const indexer = require('../../indexer');
+
 const apiOptions = key => {
   return {
     json: true,
@@ -102,44 +99,29 @@ async function scraper (key) {
     if (downloadData.thumbnail) {
       const urlBits = new URL(downloadData.thumbnail).pathname.split('/');
       const filename = urlBits[urlBits.length - 1].replace(/%20/g, '_');
-      await fs.ensureFile(`${process.env.DB_ROOT}/files/gumroad/${userId}/${product.id}/${filename}`);
-      await retry(() => {
-        return new Promise((resolve, reject) => {
-          request.get({ url: downloadData.thumbnail, encoding: null })
-            .on('complete', () => resolve())
-            .on('error', err => reject(err))
-            .pipe(fs.createWriteStream(`${process.env.DB_ROOT}/files/gumroad/${userId}/${product.id}/${filename}`));
-        });
-      });
-
+      await downloadFile({
+        ddir: path.join(process.env.DB_ROOT, `/files/gumroad/${userId}/${product.id}`),
+        name: filename
+      }, {
+        url: downloadData.thumbnail
+      })
       model.post_file.name = filename;
       model.post_file.path = `/files/gumroad/${userId}/${product.id}/${filename}`;
     }
 
     await Promise.map(downloadData.data.files, async (file) => {
-      const randomKey = crypto.randomBytes(20).toString('hex');
-      await fs.ensureFile(`${process.env.DB_ROOT}/attachments/gumroad/${userId}/${product.id}/${randomKey}`);
-      await retry(() => {
-        return new Promise((resolve, reject) => {
-          request2
-            .get('https://gumroad.com' + downloadData.data.download_info[file.id].download_url, scrapeOptions(key))
-            .on('complete', async () => {
-              const ext = file.extension.toLowerCase();
-              const filename = slugify(file.file_name, { lowercase: false });
-              model.attachments.push({
-                name: `${filename}.${ext}`,
-                path: `/attachments/gumroad/${userId}/${product.id}/${filename}.${ext}`
-              });
-              await fs.move(
-                `${process.env.DB_ROOT}/attachments/gumroad/${userId}/${product.id}/${randomKey}`,
-                `${process.env.DB_ROOT}/attachments/gumroad/${userId}/${product.id}/${filename}.${ext}`
-              );
-              resolve();
-            })
-            .on('error', err => reject(err))
-            .pipe(fs.createWriteStream(`${process.env.DB_ROOT}/attachments/gumroad/${userId}/${product.id}/${randomKey}`));
-        });
-      });
+      await downloadFile({
+        ddir: path.join(process.env.DB_ROOT, `/attachments/gumroad/${userId}/${product.id}`),
+        name: `${file.file_name}.${file.extension.toLowerCase()}`
+      }, Object.assign({
+        url: 'https://gumroad.com' + downloadData.data.download_info[file.id].download_url
+      }, scrapeOptions(key)))
+        .then(res => {
+          model.attachments.push({
+            name: res.filename,
+            path: `/attachments/gumroad/${userId}/${product.id}/${res.filename}`
+          });
+        })
     });
 
     posts.insertOne(model);